Refactor MTP logics
