+++ src/core/runner.rs
@@ -117,20 +117,19 @@ pub struct CpuTqLayerCache {
 }
 
 pub struct ModelRunner {
-    model: Model,
+    pub(crate) model: Model,
     gpu_kv_cache: Arc<Mutex<Vec<(Tensor, Tensor)>>>,
     cpu_kv_cache: Arc<Mutex<Vec<(Tensor, Tensor)>>>,
     cpu_tq_cache: Option<Vec<CpuTqLayerCache>>,
-    device: Device,
+    pub(crate) device: Device,
     config: EngineConfig,
     #[cfg(all(feature = "cuda", feature = "graph"))]
     pub decode_capturer: GraphCapturer<CudaGraphWrapper<CudaGraphFn>>,
     #[cfg(all(feature = "cuda", feature = "graph"))]
     pub mtp_capturer: Option<GraphCapturer<CudaGraphWrapper<CudaGraphFn>>>,
     #[cfg(feature = "flashinfer")]
-    flashinfer_kv_params: Option<FlashInferKvParams>,
+    pub(crate) flashinfer_kv_params: Option<FlashInferKvParams>,
     logit_processor: LogitsProcessor,
-    /// Cached sampling strategy computed once during prefill, reused during decode
     cached_sampling: RwLock<Option<CachedSamplingParams>>,
     seq_tokens: RwLock<HashMap<usize, Vec<u32>>>,
     restored_prefix_sequences: RwLock<HashSet<usize>>,
@@ -139,19 +138,12 @@ pub struct ModelRunner {
     guidance_mismatch: RwLock<HashSet<usize>>,
     llg_factory: Option<Arc<ParserFactory>>,
     transfer: Option<Arc<Transfer>>,
-    /// Whether this runner is on the first rank (for logging)
     is_first_rank: bool,
-    model_type: ModelType,
+    pub(crate) model_type: ModelType,
     /// MTP head for speculative decoding (Qwen3.5 only for now)
-    mtp_head: Option<Arc<Qwen3_5MtpHead>>,
+    pub(crate) mtp_head: Option<Arc<Qwen3_5MtpHead>>,
     /// Number of speculative tokens to draft per step
-    mtp_num_speculative: usize,
-}
-
-struct MtpSeqInfo {
-    id: usize,
-    len: usize,
-    block_table: Vec<u32>,
+    pub(crate) mtp_num_speculative: usize,
 }
 
 impl ModelRunner {
@@ -160,144 +152,31 @@ impl ModelRunner {
     #[cfg(all(feature = "cuda", feature = "graph"))]
     const GRAPH_CAPTURE_MIN_BATCH: usize = 16;
 
-    fn is_mla_model(&self) -> bool {
+    pub(crate) fn is_mla_model(&self) -> bool {
         matches!(
             self.model_type,
             ModelType::GLM4MoeLite | ModelType::DeepSeek
         )
     }
 
-    fn compute_slot_mappings(
-        &self,
-        seq_info: &MtpSeqInfo,
-        num_tokens: usize,
-        block_size: usize,
-        ctx: &str,
-    ) -> Result<Vec<i64>> {
-        let mut slots = Vec::with_capacity(num_tokens);
-        for i in 0..num_tokens {
-            let pos = seq_info.len + i;
-            let block_idx = pos / block_size;
-            let block_offset = pos % block_size;
-            if block_idx < seq_info.block_table.len() {
-                let physical_block = seq_info.block_table[block_idx] as i64;
-                slots.push(physical_block * block_size as i64 + block_offset as i64);
-            } else {
-                candle_core::bail!(
-                    "MTP {} missing KV block: block_idx {} >= block_table.len() {}. \
-                     Blocks must be pre-allocated before MTP.",
-                    ctx,
-                    block_idx,
-                    seq_info.block_table.len()
-                );
-            }
-        }
-        Ok(slots)
+    pub(crate) fn model(&self) -> &Model {
+        &self.model
     }
 
-    fn build_mtp_metadata(
-        &self,
-        seq_info: &MtpSeqInfo,
-        slot_mappings: &[i64],
-        q_len: usize,
-    ) -> Result<InputMetadata> {
-        let total_kv_len = (seq_info.len + q_len) as u32;
-        let mamba_slot_mapping = self.prepare_mamba_slot_mapping(&[seq_info.id], false)?;
-
-        #[cfg(feature = "flashinfer")]
-        let flashinfer_metadata = if let Some(params) = self.flashinfer_kv_params {
-            let num_pages = seq_info.block_table.len();
-            let indptr_host = vec![0u32, num_pages as u32];
-            let indices_vec: Vec<u32> = seq_info.block_table.clone();
-            let last_page_tokens =
-                total_kv_len as usize - (num_pages.saturating_sub(1)) * params.page_size;
-            let last_len_host = vec![last_page_tokens as u32];
-            let kv_len_arr_host = vec![total_kv_len];
-            let q_cu_seqlens_host = vec![0u32, q_len as u32];
-
-            #[cfg(all(feature = "cuda", feature = "graph"))]
-            let use_graph = self
-                .mtp_capturer
-                .as_ref()
-                .map_or(false, |c| c.is_mtp_captured(q_len));
-            #[cfg(not(all(feature = "cuda", feature = "graph")))]
-            let use_graph = false;
-
-            let prefill_plan_info = if use_graph {
-                None
-            } else {
-                Some(attention_rs::flashinfer::prefill_plan(
-                    &self.device,
-                    &q_cu_seqlens_host,
-                    &indptr_host,
-                    &kv_len_arr_host,
-                    q_len as u32,
-                    1,
-                    params.num_qo_heads,
-                    params.num_kv_heads,
-                    params.head_dim,
-                    params.page_size,
-                    params.out_dtype,
-                    None,
-                    Some(params.kv_dtype),
-                    false,
-                )?)
-            };
-
-            Some(attention_rs::FlashInferMetadata {
-                indptr: Tensor::from_vec(indptr_host.clone(), (2,), &self.device)?,
-                indptr_host,
-                indices: Tensor::from_vec(indices_vec, (num_pages,), &self.device)?,
-                last_len: Tensor::from_vec(last_len_host.clone(), (1,), &self.device)?,
-                last_len_host: Some(last_len_host),
-                kv_len_arr_host: Some(kv_len_arr_host),
-                total_num_rows: Some(q_len as u32),
-                batch_indices: None,
-                positions: None,
-                use_cuda_graph: use_graph,
-                decode_plan_info: None,
-                prefill_plan_info,
-                mla_decode_plan_info: None,
-                mla_prefill_plan_info: None,
-            })
-        } else {
-            None
-        };
-        #[cfg(not(feature = "flashinfer"))]
-        let flashinfer_metadata = None;
-
-        Ok(InputMetadata {
-            is_prefill: true,
-            is_mla: self.is_mla_model(),
-            sequence_ids: Some(vec![seq_info.id]),
-            mamba_slot_mapping,
-            slot_mapping: Tensor::from_vec(slot_mappings.to_vec(), (q_len,), &self.device)?,
-            context_lens: Some(Tensor::from_vec(vec![total_kv_len], (1,), &self.device)?),
-            block_tables: Some(Tensor::from_vec(
-                seq_info.block_table.clone(),
-                (1, seq_info.block_table.len()),
-                &self.device,
-            )?),
-            seqlens: None,
-            cu_seqlens_q: Some(Tensor::from_vec(
-                vec![0u32, q_len as u32],
-                (2,),
-                &self.device,
-            )?),
-            cu_seqlens_k: Some(Tensor::from_vec(
-                vec![0u32, total_kv_len],
-                (2,),
-                &self.device,
-            )?),
-            max_seqlen_q: q_len,
-            max_seqlen_k: seq_info.len + q_len,
-            max_context_len: seq_info.len + q_len,
-            flashinfer_metadata,
-            is_mtp_verify: true,
-        })
+    pub(crate) fn device(&self) -> &Device {
+        &self.device
     }
 
-    fn prepare_mamba_slot_mapping(
+    pub(crate) fn block_size(&self) -> usize {
+        self.config.block_size
+    }
+
+    #[cfg(feature = "flashinfer")]
+    pub(crate) fn flashinfer_kv_params(&self) -> Option<crate::utils::FlashInferKvParams> {
+        self.flashinfer_kv_params
+    }
+
+    pub(crate) fn prepare_mamba_slot_mapping(
         &self,
         sequence_ids: &[usize],
         is_prefill: bool,
@@ -1122,15 +1001,6 @@ impl ModelRunner {
         }
     }
 
-    fn mtp_rollback_mamba(&self, seq_id: usize, keep_tokens: usize) -> Result<bool> {
-        match &self.model {
-            Model::Qwen3_5(m) => m.mtp_rollback_mamba(seq_id, keep_tokens),
-            Model::Qwen3_5MoE(m) => m.mtp_rollback_mamba(seq_id, keep_tokens),
-            Model::Qwen3VL(m) => m.mtp_rollback_mamba(seq_id, keep_tokens),
-            _ => Ok(false),
-        }
-    }
-
     #[allow(unused)]
     pub fn run(&self, seqs: Seqs, is_prefill: bool) -> Result<Vec<u32>> {
         #[cfg(feature = "nvtx")]
@@ -1285,365 +1155,6 @@ impl ModelRunner {
         #[cfg(feature = "nvtx")]
         nvtx::range_pop!();
         Ok(output_ids)
-    }
-
-    /// MTP Step 1: single-token decode to get anchor token + hidden state.
-    /// Tries CUDA graph replay first (the graph's internal buffer for the
-    /// post-norm hidden state is accessible via take_last_hidden_for_mtp),
-    /// falling back to eager forward_with_hidden.
-    fn mtp_decode_step1(&self, seqs: Seqs, _seq_info: &MtpSeqInfo) -> Result<(u32, Tensor)> {
-        let (input_ids, positions, mut input_metadata) = match &seqs {
-            Seqs::SeqRefs(seqs_ref) => self.prepare_decode(*seqs_ref)?,
-            Seqs::DecodeVec(decode_seqs) => self.prepare_decode(decode_seqs.iter())?,
-        };
-
-        let _decode_guard = set_linear_is_prefill(false);
-
-        // Try CUDA graph replay for the decode forward. The model's forward()
-        // stores hidden states in last_hidden_for_mtp during both capture and
-        // replay (the cached tensor shares GPU storage with the graph output,
-        // so it's updated in-place on replay).
-        #[cfg(all(feature = "cuda", feature = "graph"))]
-        {
-            let input_batch = input_ids.dim(0)?;
-            let require_exact_graph = input_metadata.mamba_slot_mapping.is_some();
-            let can_replay = if require_exact_graph {
-                self.decode_capturer.is_exact_captured(input_batch)
-            } else {
-                self.decode_capturer.is_captured(input_batch)
-            };
-            if can_replay {
-                let logits = match &self.model {
-                    Model::Qwen3_5(model) => {
-                        let _guard = model.lock_mamba_cache_for_graph();
-                        self.decode_capturer
-                            .replay(&input_ids, &positions, &input_metadata)?
-                    }
-                    Model::Qwen3_5MoE(model) => {
-                        let _guard = model.lock_mamba_cache_for_graph();
-                        self.decode_capturer
-                            .replay(&input_ids, &positions, &input_metadata)?
-                    }
-                    Model::Qwen3VL(model) => {
-                        if let Some(_guard) = model.lock_mamba_cache_for_graph() {
-                            self.decode_capturer
-                                .replay(&input_ids, &positions, &input_metadata)?
-                        } else {
-                            self.decode_capturer
-                                .replay(&input_ids, &positions, &input_metadata)?
-                        }
-                    }
-                    _ => self
-                        .decode_capturer
-                        .replay(&input_ids, &positions, &input_metadata)?,
-                };
-
-                let hidden_states = match &self.model {
-                    Model::Qwen3_5(model) => model.take_last_hidden_for_mtp(),
-                    Model::Qwen3_5MoE(model) => model.take_last_hidden_for_mtp(),
-                    Model::Qwen3VL(model) => model.take_last_hidden_for_mtp(),
-                    _ => None,
-                };
-
-                if let Some(hidden_states) = hidden_states {
-                    let anchor_token = self.sample(&logits, seqs, false)?[0];
-                    let seq_hidden = if hidden_states.dims().len() == 2 && hidden_states.dim(0)? > 1
-                    {
-                        hidden_states.get(hidden_states.dim(0)? - 1)?
-                    } else if hidden_states.dims().len() == 2 {
-                        hidden_states.get(0)?
-                    } else {
-                        hidden_states
-                    };
-                    return Ok((anchor_token, seq_hidden));
-                }
-            }
-        }
-
-        // Fallback: eager forward_with_hidden (no graph available or hidden state extraction failed)
-        #[cfg(feature = "flashinfer")]
-        if let Some(fm) = input_metadata.flashinfer_metadata.as_mut() {
-            if input_metadata.is_mla {
-                if fm.mla_decode_plan_info.is_none() {
-                    if let Some(params) = self.flashinfer_kv_params {
-                        fm.mla_decode_plan_info = Some(attention_rs::mla::mla_decode_plan(
-                            &self.device,
-                            params.kv_dtype,
-                            &fm.indptr_host,
-                            input_ids.dim(0)?,
-                            params.num_qo_heads,
-                            params.page_size,
-                            fm.use_cuda_graph,
-                        )?);
-                    }
-                }
-            } else if fm.decode_plan_info.is_none() {
-                if let Some(params) = self.flashinfer_kv_params {
-                    fm.decode_plan_info = Some(attention_rs::flashinfer::decode_plan(
-                        &self.device,
-                        params.kv_dtype,
-                        params.out_dtype,
-                        &fm.indptr_host,
-                        fm.last_len_host.as_deref(),
-                        fm.kv_len_arr_host.as_deref(),
-                        input_ids.dim(0)?,
-                        params.num_qo_heads,
-                        params.num_kv_heads,
-                        params.head_dim,
-                        params.page_size,
-                        fm.use_cuda_graph,
-                    )?);
-                }
-            }
-        }
-
-        let kv_cache = self.get_kv_cache();
-        let (logits, hidden_states) = match &self.model {
-            Model::Qwen3_5(model) => model.forward_with_hidden(
-                &input_ids,
-                &positions,
-                Some(&kv_cache),
-                &input_metadata,
-                false,
-            )?,
-            Model::Qwen3_5MoE(model) => model.forward_with_hidden(
-                &input_ids,
-                &positions,
-                Some(&kv_cache),
-                &input_metadata,
-                false,
-            )?,
-            Model::Qwen3VL(model) => model.forward_with_hidden(
-                &input_ids,
-                &positions,
-                Some(&kv_cache),
-                &input_metadata,
-                false,
-            )?,
-            _ => {
-                drop(kv_cache);
-                candle_core::bail!("MTP Step 1 requires Qwen3.5 model");
-            }
-        };
-        drop(kv_cache);
-
-        let anchor_token = self.sample(&logits, seqs, false)?[0];
-
-        let seq_hidden = if hidden_states.dims().len() == 2 && hidden_states.dim(0)? > 1 {
-            hidden_states.get(hidden_states.dim(0)? - 1)?
-        } else if hidden_states.dims().len() == 2 {
-            hidden_states.get(0)?
-        } else {
-            hidden_states.clone()
-        };
-
-        Ok((anchor_token, seq_hidden))
-    }
-
-    /// Run MTP speculative decode for a batch of sequences.
-    /// Returns Vec<Vec<u32>> where each inner vec contains all accepted tokens for that sequence
-    /// (anchor + accepted drafts + bonus token).
-    ///
-    /// Optimized flow:
-    ///   1. Run main model decode via CUDA graph replay (when available) + extract hidden state
-    ///   2. Sample anchor token from logits
-    ///   3. MTP head drafts K tokens autoregressively (no KV cache)
-    ///   4. Verify: run main model on [anchor, draft_0, ..., draft_{K-1}] using native flash
-    ///   5. On partial rejection: roll back GDN state to the accepted token boundary
-    ///   6. Greedy-accept matching prefix; take bonus token at first mismatch
-    pub fn run_mtp_decode(&self, seqs: Seqs) -> Result<Vec<Vec<u32>>> {
-        let mtp_head = match &self.mtp_head {
-            Some(h) => h.clone(),
-            None => {
-                let output = self.run(seqs, false)?;
-                return Ok(output.into_iter().map(|t| vec![t]).collect());
-            }
-        };
-
-        let (batch_size, seq_infos) = match &seqs {
-            Seqs::SeqRefs(s) => {
-                let infos: Vec<MtpSeqInfo> = s
-                    .iter()
-                    .map(|seq| MtpSeqInfo {
-                        id: seq.id,
-                        len: seq.len(),
-                        block_table: seq.block_table.clone(),
-                    })
-                    .collect();
-                (s.len(), infos)
-            }
-            Seqs::DecodeVec(d) => {
-                let infos: Vec<MtpSeqInfo> = d
-                    .iter()
-                    .map(|ds| MtpSeqInfo {
-                        id: ds.id,
-                        len: ds.len,
-                        block_table: ds.block_tables.clone(),
-                    })
-                    .collect();
-                (d.len(), infos)
-            }
-        };
-
-        if batch_size != 1 {
-            let output = self.run(seqs, false)?;
-            return Ok(output.into_iter().map(|t| vec![t]).collect());
-        }
-
-        let seq_info = &seq_infos[0];
-        let num_draft = self.mtp_num_speculative;
-
-        // Step 1: Main model decode for logits + hidden state.
-        let (anchor_token, seq_hidden) = self.mtp_decode_step1(seqs, seq_info)?;
-
-        // Step 2: Draft K tokens using MTP head (GPU-resident, no per-step CPU sync)
-        let embed_weight = match &self.model {
-            Model::Qwen3_5(m) => m.embed_weight().clone(),
-            Model::Qwen3_5MoE(m) => m.embed_weight().clone(),
-            Model::Qwen3VL(m) => m
-                .embed_weight()
-                .expect("Qwen3VL MTP requires Qwen3.5 text backbone")
-                .clone(),
-            _ => unreachable!(),
-        };
-        let lm_head_fn = |hidden: &Tensor| -> Result<Tensor> {
-            match &self.model {
-                Model::Qwen3_5(m) => m.forward_lm_head(hidden),
-                Model::Qwen3_5MoE(m) => m.forward_lm_head(hidden),
-                Model::Qwen3VL(m) => m.forward_lm_head(hidden),
-                _ => unreachable!(),
-            }
-        };
-
-        let base_position = seq_info.len.saturating_sub(1);
-        let anchor_token_tensor = Tensor::from_vec(vec![anchor_token], (1,), &self.device)?;
-        let (draft_tokens, _last_hidden) = mtp_head.draft_tokens_gpu(
-            &seq_hidden,
-            &anchor_token_tensor,
-            num_draft,
-            &embed_weight,
-            lm_head_fn,
-            base_position,
-        )?;
-
-        if draft_tokens.is_empty() {
-            return Ok(vec![vec![anchor_token]]);
-        }
-
-        // Step 3: Verify draft tokens via prefill-style forward on [anchor, draft_0..K-1].
-        let mut verify_tokens = vec![anchor_token];
-        verify_tokens.extend_from_slice(&draft_tokens);
-        let verify_len = verify_tokens.len();
-
-        let block_size = self.config.block_size;
-        let slot_mappings =
-            self.compute_slot_mappings(seq_info, verify_len, block_size, "verify")?;
-
-        let verify_input_ids = Tensor::from_vec(verify_tokens, (verify_len,), &self.device)?;
-        let verify_positions_tensor = Tensor::from_vec(
-            (0..verify_len)
-                .map(|i| (seq_info.len + i) as i64)
-                .collect::<Vec<_>>(),
-            (verify_len,),
-            &self.device,
-        )?;
-
-        let verify_metadata =
-            self.build_mtp_metadata(seq_info, &slot_mappings[..verify_len], verify_len)?;
-
-        let _prefill_guard = set_linear_is_prefill(true);
-
-        #[cfg(all(feature = "cuda", feature = "graph"))]
-        let use_mtp_graph = self
-            .mtp_capturer
-            .as_ref()
-            .map_or(false, |c| c.is_mtp_captured(verify_len));
-        #[cfg(not(all(feature = "cuda", feature = "graph")))]
-        let use_mtp_graph = false;
-
-        let all_logits_result = if use_mtp_graph {
-            #[cfg(all(feature = "cuda", feature = "graph"))]
-            {
-                self.mtp_capturer.as_ref().unwrap().replay_mtp(
-                    &verify_input_ids,
-                    &verify_positions_tensor,
-                    &verify_metadata,
-                )
-            }
-            #[cfg(not(all(feature = "cuda", feature = "graph")))]
-            {
-                unreachable!()
-            }
-        } else {
-            let kv_cache = self.get_kv_cache();
-            let res = match &self.model {
-                Model::Qwen3_5(model) => model.forward(
-                    &verify_input_ids,
-                    &verify_positions_tensor,
-                    Some(&kv_cache),
-                    &verify_metadata,
-                    false,
-                ),
-                Model::Qwen3_5MoE(model) => model.forward(
-                    &verify_input_ids,
-                    &verify_positions_tensor,
-                    Some(&kv_cache),
-                    &verify_metadata,
-                    false,
-                ),
-                Model::Qwen3VL(model) => model.forward(
-                    &verify_input_ids,
-                    &verify_positions_tensor,
-                    Some(&kv_cache),
-                    &verify_metadata,
-                    None,
-                ),
-                _ => unreachable!(),
-            };
-            drop(kv_cache);
-            res
-        };
-        let all_logits = match all_logits_result {
-            Ok(logits) => logits,
-            Err(err) => {
-                return Err(err);
-            }
-        };
-
-        let verify_result = match crate::core::mtp::verify_draft_greedy(&all_logits, &draft_tokens)
-        {
-            Ok(result) => result,
-            Err(err) => {
-                return Err(err);
-            }
-        };
-
-        if verify_result.num_accepted < verify_result.num_proposed {
-            let commit_len = 1 + verify_result.num_accepted;
-            // KV cache does not need explicit rollback because the next decode writes
-            // the continuation token at the first rejected slot and stale later slots
-            // are outside the sequence length.
-            let restored = self.mtp_rollback_mamba(seq_info.id, commit_len)?;
-            if !restored {
-                candle_core::bail!(
-                    "MTP failed to roll back mamba-state snapshot for seq {} to {} verified token(s)",
-                    seq_info.id,
-                    commit_len
-                );
-            }
-        }
-
-        let mut result_tokens = Vec::with_capacity(2 + verify_result.num_accepted);
-        result_tokens.push(anchor_token);
-        result_tokens.extend_from_slice(&verify_result.accepted_tokens);
-        result_tokens.push(verify_result.continuation_token);
-
-        crate::core::mtp::mtp_stats_update(verify_result.num_proposed, verify_result.num_accepted);
-        if crate::core::mtp::MTP_TOTAL_STEPS.load(std::sync::atomic::Ordering::Relaxed) % 256 == 0 {
-            crate::log_info!("{}", crate::core::mtp::mtp_stats_summary());
-        }
-
-        Ok(vec![result_tokens])
     }
 
     pub fn embed(&self, seqs: &[&Sequence], strategy: &EmbeddingStrategy) -> Result<Vec<Vec<f32>>> {
@@ -1981,7 +1492,10 @@ impl ModelRunner {
         Ok((input_ids, positions, input_metadata))
     }
 
-    fn prepare_decode<'a, I, S>(&self, seqs: I) -> Result<(Tensor, Tensor, InputMetadata)>
+    pub(crate) fn prepare_decode<'a, I, S>(
+        &self,
+        seqs: I,
+    ) -> Result<(Tensor, Tensor, InputMetadata)>
     where
         I: IntoIterator<Item = &'a S>,
         S: ToDecodeInput + 'a,
@@ -2131,7 +1645,7 @@ impl ModelRunner {
         Ok((input_ids, positions, input_metadata))
     }
 
-    fn sample(&self, logits: &Tensor, seqs: Seqs, is_prefill: bool) -> Result<Vec<u32>> {
+    pub(crate) fn sample(&self, logits: &Tensor, seqs: Seqs, is_prefill: bool) -> Result<Vec<u32>> {
         let seq_ids: Vec<usize> = match &seqs {
             Seqs::SeqRefs(seqs) => seqs.iter().map(|s| s.id()).collect(),
             Seqs::DecodeVec(v) => v.iter().map(|s| s.id()).collect(),
